rasberry_pi_code/requirements.txt
#!/usr/bin/env python3
"""
Requirements file for Plant Monitoring System
"""

paho-mqtt==1.6.1
adafruit-circuitpython-dht==3.7.8
board>=1.0
smbus2==0.4.2
pyserial==3.5
python-dotenv==1.0.0